update CHANGELOG and README
