+++ CHANGELOG.md
@@ -1,0 +1,17 @@
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+- [Add DIDcomm Service]()
+- [Add keychain test]()
+- [Add EtherIdentityService and WalletService Test]()
+- [Add config and Fix Lint ]()
+- [Add function returns]()
+- [Add documentation]()
+- [Added class diagram]()
+- [Add Presentation Service]()
+- [Add Credential Service]()
+- [Add General Identity Service]()
